+++ src/BenchmarksApps/TechEmpower/BlazorSSR/BlazorSSR.csproj
@@ -11,6 +11,5 @@
   <ItemGroup>
     <PackageReference Include="Dapper" Version="2.0.123" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.0-preview.4" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.0-preview.4.23259.3" />
   </ItemGroup>
 </Project>
